Strip out the transactional interfaces
